test(opencode): split model fallback coverage
